I finally defeated every avoidable warning (#335)
* silver coin labels

* match obj_loop_log

* better

* match obj_loop_modechange

* nonmatching obj_loop_fireball_octoweapon

* fix racer.c warnings

* match obj_loop_fireball_octoweapon

* Update object_functions.c

* match obj_loop_lasergun

* unknown_0255E0.c warnings fixed

* asset_loading.c warnings too

* laser gun labels

* quick fixes

* match func_8003F0F8

* prelim

* match func_8000E4E8

* warnings are dead

* quick fixes

* Update math_util.c

* reee

+++ src/gzip.c
@@ -42,7 +42,7 @@ s32 byteswap32(u8 *arg0) {
  * Returns the uncompressed size of a gzip compressed asset.
  */
 s32 get_asset_uncompressed_size(s32 assetIndex, s32 assetOffset) {
-    load_asset_to_address(assetIndex, gAssetAddress, assetOffset, 8);
+    load_asset_to_address(assetIndex, (u32) gAssetAddress, assetOffset, 8);
     return byteswap32((u8 *)gAssetAddress);
 }
